# services/order-management/openapi.yaml
...
openapi: 3.0.3
info:
  title: Waypoint Order Management API
  version: 1.0.0
  description: >-
    Order lifecycle for the Waypoint delivery logistics platform: drafts, the 16:00 Asia/Colombo cutoff,

    deferrals, Planning write-back, store receipts and the dispatcher dashboard.


    ### Authentication

    Every route except `/health` needs the header `Authorization: Bearer <access_token>`.

    Get a token from auth-rbac (through the gateway your environment uses), then press **Authorize** and paste it

    (without the `Bearer ` prefix):


    ```bash

    curl -X POST "$GATEWAY_URL/api/auth/login" \
      -H 'Content-Type: application/json' \
      -d '{"username":"manager_out001","password":"Password123!"}'
    ```


    Seeded accounts (password `Password123!`): `dispatcher_admin`, `manager_out001` (outlet OUT001),

    `loader_peliyagoda`, `driver_colombo`.


    ### Headers

    | Header | Where | Purpose |

    |---|---|---|

    | `Authorization: Bearer <token>` | all but `/health` | access token from auth-rbac |

    | `Content-Type: application/json` | requests with a body | |

    | `Idempotency-Key` | `POST /` | safe retries of create |

    | `X-Test-Now` | any (non-production) | freeze the business clock for one request |


    ### Responses

    Success: `{ "success": true, "data": … }`. Failure: `{ "success": false, "error": { code, message, details? } }`.

    Dates are `YYYY-MM-DD`, delivery windows `HH:mm` (Asia/Colombo), instants ISO 8601 UTC.


...

...

...

...


...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from `POST /api/auth/login`. Sent as `Authorization: Bearer <token>`."
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: CUTOFF_PASSED
            message:
              type: string
            details:
              description: Field list, next_available_date, failures[] etc.
    OrderStatus:
      type: string
      enum:
        - draft
        - confirmed
        - allocated
        - loaded
        - out_for_delivery
        - delivered
        - received
        - disputed
        - deferred
        - not_run
        - cancelled
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_ref:
          type: string
        sku:
          type: string
          example: MLK-1L
        description:
          type: string
        quantity:
          type: integer
        unit_weight_kg:
          type: number
        unit_volume_m3:
          type: number
        is_chilled:
          type: boolean
        created_at:
          type: string
          format: date-time
    OrderEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_ref:
          type: string
        from_status:
          type: string
          nullable: true
        to_status:
          $ref: "#/components/schemas/OrderStatus"
        reason_code:
          type: string
          nullable: true
        reason_note:
          type: string
          nullable: true
        actor_id:
          type: string
          nullable: true
        actor_role:
          type: string
        occurred_at:
          type: string
          format: date-time
    Order:
      type: object
      properties:
        order_ref:
          type: string
          example: ORD-20260929-00042
        outlet_id:
          type: string
          example: OUT001
        brand:
          type: string
          enum:
            - Fresh
            - Style
            - Tech
        depot:
          type: string
          enum:
            - Peliyagoda
            - Kandy
        order_date:
          type: string
          format: date
        original_order_date:
          type: string
          format: date
        requested_order_date:
          type: string
          format: date
          nullable: true
        temp_requirement:
          type: string
          enum:
            - ambient
            - chilled
        status:
          $ref: "#/components/schemas/OrderStatus"
        order_units:
          type: integer
        order_weight_kg:
          type: number
        order_volume_m3:
          type: number
        window_open_time:
          type: string
          example: 05:00
        window_close_time:
          type: string
          example: 07:30
        placed_by:
          type: string
          nullable: true
        placed_by_username:
          type: string
        placed_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
        cutoff_applied_at:
          type: string
          format: date-time
          nullable: true
        deferral_count:
          type: integer
        vehicle_id:
          type: string
          nullable: true
        trip_id:
          type: integer
          nullable: true
        idempotency_key:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            events:
              type: array
              description: Latest 20, oldest first
              items:
                $ref: "#/components/schemas/OrderEvent"
paths:
...
  /{order_ref}/confirm:
    post:
      operationId: confirmOrder
      tags:
        - Orders
      summary: Confirm a draft under the 16:00 cutoff
      description: "Roles: store_manager (own outlet), dispatcher. Before 16:00 Asia/Colombo the order is confirmed for the
        next operating day. At or after 16:00 it returns 409 CUTOFF_PASSED unless `accept_next_run` is true, in which
        case it rolls to the following run. Use `X-Test-Now` to try both sides of the cutoff."
      security: *a1
      parameters:
        - *a8
        - *a10
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                accept_next_run:
                  type: boolean
                  default: false
              additionalProperties: false
            examples:
              beforeCutoff:
                summary: Normal confirm (no body needed)
                value: {}
              acceptNextRun:
                summary: After cutoff — accept the following run
                value:
                  accept_next_run: true
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    allOf:
                      - $ref: "#/components/schemas/OrderDetail"
                      - type: object
                        properties:
                          rolled_to_next_run:
                            type: boolean
              example:
                success: true
                data:
                  order_ref: ORD-20260929-00042
                  outlet_id: OUT001
                  brand: Fresh
                  depot: Peliyagoda
                  order_date: 2026-09-29
                  original_order_date: 2026-09-29
                  requested_order_date: null
                  temp_requirement: chilled
                  status: confirmed
                  order_units: 120
                  order_weight_kg: 123.6
                  order_volume_m3: 0.132
                  window_open_time: 05:00
                  window_close_time: 07:30
                  placed_by: 9595b3cf-eec1-4559-9fcc-889ef5a9280c
                  placed_by_username: manager_out001
                  placed_at: 2026-09-28T04:30:00.000Z
                  confirmed_at: 2026-09-28T04:35:00.000Z
                  cutoff_applied_at: 2026-09-28T04:35:00.000Z
                  deferral_count: 0
                  vehicle_id: null
                  trip_id: null
                  idempotency_key: null
                  created_at: 2026-09-28T04:30:00.412Z
                  updated_at: 2026-09-28T04:35:00.108Z
                  items: *a16
                  events: *a17
                  rolled_to_next_run: false
        "401": *a4
        "403":
          description: FORBIDDEN · OUTLET_SCOPE_VIOLATION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    success: false
                    error:
                      code: FORBIDDEN
                      message: "Access denied. Requires one of roles: [store_manager, dispatcher]"
                OUTLET_SCOPE_VIOLATION:
                  summary: OUTLET_SCOPE_VIOLATION
                  value:
                    success: false
                    error:
                      code: OUTLET_SCOPE_VIOLATION
                      message: Store manager for OUT001 may not access outlet OUT050
                      details:
                        token_outlet_id: OUT001
                        requested_outlet_id: OUT050
        "404": *a9
        "409":
          description: CUTOFF_PASSED · INVALID_STATE_TRANSITION · DUPLICATE_ORDER
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                CUTOFF_PASSED:
                  summary: CUTOFF_PASSED
                  value:
                    success: false
                    error:
                      code: CUTOFF_PASSED
                      message: 'The 16:00 Asia/Colombo cutoff for 2026-09-29 has passed. Resend with {"accept_next_run": true} to confirm for
                        the next available run.'
                      details:
                        requested_date: 2026-09-29
                        next_available_date: 2026-09-30
                INVALID_STATE_TRANSITION:
                  summary: INVALID_STATE_TRANSITION
                  value:
                    success: false
                    error:
                      code: INVALID_STATE_TRANSITION
                      message: Only draft orders can be confirmed; ORD-20260929-00042 is 'confirmed'
                      details:
                        from: confirmed
                        to: confirmed
                        allowed:
                          - allocated
                          - deferred
                          - cancelled
                DUPLICATE_ORDER:
                  summary: DUPLICATE_ORDER
                  value:
                    success: false
                    error:
                      code: DUPLICATE_ORDER
                      message: Outlet OUT001 already has an active chilled order for 2026-09-29
                      details:
                        existing_order_ref: ORD-20260929-00040
                        outlet_id: OUT001
                        order_date: 2026-09-29
                        temp_requirement: chilled
        "422":
          description: EMPTY_ORDER · NON_OPERATING_DATE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                EMPTY_ORDER:
                  summary: EMPTY_ORDER
                  value:
                    success: false
                    error:
                      code: EMPTY_ORDER
                      message: Add at least one item before confirming
                      details:
                        order_ref: ORD-20260929-00042
                NON_OPERATING_DATE:
                  summary: NON_OPERATING_DATE
                  value:
                    success: false
                    error:
                      code: NON_OPERATING_DATE
                      message: 2026-05-01 is no longer an operating day
                      details:
                        order_date: 2026-05-01
                        next_operating_day: 2026-05-02
        "503": *a5
...